+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -27,7 +27,6 @@ use Symfony\Component\JsonStreamer\JsonStreamWriter;
 use Symfony\Component\Mailer\Mailer;
 use Symfony\Component\Notifier\Notifier;
 use Symfony\Component\RateLimiter\Policy\TokenBucketLimiter;
-use Symfony\Component\Scheduler\Messenger\SchedulerTransportFactory;
 use Symfony\Component\Serializer\Encoder\JsonDecode;
 use Symfony\Component\Webhook\Controller\WebhookController;
 
@@ -920,9 +919,6 @@ class ConfigurationTest extends TestCase
                     'cache_pool' => 'cache.rate_limiter',
                     'storage_service' => null,
                 ],
-            ],
-            'scheduler' => [
-                'enabled' => !class_exists(FullStack::class) && class_exists(SchedulerTransportFactory::class),
             ],
             'exceptions' => [],
             'webhook' => [