+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -22,6 +22,8 @@ CHANGELOG
  * Register `PropertyAccessBundle` automatically when the PropertyAccess component is installed
  * Make `framework.uid` an alias of the `uid` configuration provided by `UidBundle`
  * Register `UidBundle` automatically when the Uid component is installed
+ * Make `framework.scheduler` an alias of the `scheduler` configuration provided by `SchedulerBundle`
+ * Register `SchedulerBundle` automatically when the Scheduler component is installed
  * Make `framework.workflows` an alias of the `workflow` configuration provided by `WorkflowBundle`
  * Register `WorkflowBundle` automatically when the Workflow component is installed
  * Make `framework.remote_event` an alias of the `remote_event` configuration provided by `RemoteEventBundle`